+++ src/main/kotlin/lib/graphics/turtle/awt/Utils.kt
@@ -2,7 +2,7 @@ package lib.graphics.turtle.awt
 
 import java.time.LocalDateTime
 
-fun log(msg: String): Unit {
+fun log(msg: String) {
     val thread = Thread.currentThread()
     val now = LocalDateTime.now()
     println("$now $thread $msg")
